utils.py splitted
