+++ src/utils/read_params.py
@@ -1,0 +1,11 @@
+import yaml
+
+
+def read_params():
+    with open("params.yaml", 'r') as fd:
+        params = yaml.safe_load(fd)
+        return params
+
+
+if __name__ == '__main__':
+    pass
